chatbot/requirements.txt
streamlit==1.28.0
fastapi==0.104.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.346
beautifulsoup4==4.12.2
uvicorn==0.23.2
transformers==4.34.0